mint_coast/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.5
defusedxml==0.8.0rc2
Django==4.2.5
django-allauth==0.57.0
django-filter==23.3
django-recaptcha3==0.4.0
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.2
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==2.0.5
